+++ files/chef-marketplace-cookbooks/chef-marketplace/recipes/_reckoner_disable.rb
@@ -1,0 +1,3 @@
+runit_service 'reckoner' do
+  action :disable
+end