+++ files/chef-marketplace-cookbooks/chef-marketplace/recipes/_server_enable.rb
@@ -28,27 +28,27 @@ template '/etc/cron.d/reporting-partition-cleanup' do
 end
 
 directory '/etc/opscode' do
-  owner 'root'
-  group 'root'
+  owner 'opscode'
+  group 'opscode'
   action :create
 end
 
 template '/etc/opscode/chef-server.rb' do
-  owner 'root'
-  group 'root'
+  source 'chef-server.rb.erb'
+  owner 'opscode'
+  group 'opscode'
   action :create_if_missing
 end
 
 directory '/etc/chef-manage' do
-  owner 'root'
-  group 'root'
+  owner 'opscode'
+  group 'opscode'
   action :create
 end
 
-file '/etc/chef-manage/manage.rb' do
-  owner 'root'
-  group 'root'
-  # TODO: enable the marketplace sign_up after it's released
-  #  content 'marketplace.sign_up.enabled = true'
-  action :create
+template '/etc/chef-manage/manage.rb' do
+  source 'manage.rb.erb'
+  owner 'opscode'
+  group 'opscode'
+  action :create_if_missing
 end