+++ files/chef-marketplace-cookbooks/chef-marketplace/recipes/enable.rb
@@ -1,3 +1,4 @@
+# Configure the node attributes
 include_recipe 'chef-marketplace::config'
 
 # Setup the MOTD first so that the user doesn't see old data if the shell in
@@ -9,6 +10,7 @@ motd '50-chef-marketplace-appliance' do
   action motd_action
 end
 
+# Register as a Chef Server plugin
 include_recipe 'chef-marketplace::_register_plugin'
 
 # 'server', 'analytics', 'aio', 'compliance'
@@ -17,8 +19,20 @@ role = node['chef-marketplace']['role']
 # Base recipes for role
 include_recipe "chef-marketplace::_#{role}_enable"
 
-# Setup billing daemon
-include_recipe 'chef-marketplace::_reckoner'
+# Configure runit
+include_recipe 'chef-marketplace::_runit'
+
+# Enable/Disable chef-marketplace services
+%w(
+  reckoner
+  biscotti
+).each do |service|
+  if node['chef-marketplace'][service]['enabled']
+    include_recipe "chef-marketplace::_#{service}_enable"
+  else
+    include_recipe "chef-marketplace::_#{service}_disable"
+  end
+end
 
 # Setup omnibus-ctl commands
 include_recipe 'chef-marketplace::_omnibus_commands'